updates for appimage creation

+++ quake-scripts/ezquake-generate-appimage/manual_creation.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
+
+SKIP_BUILD=0
+
 if ! hash appimagetool;then
 	echo "appimagetool not installed."
+	exit 1
+fi
+if ! hash patchelf;then
+	echo "patchelf not installed."
 	exit 1
 fi
 
@@ -18,6 +25,9 @@ export LDFLAGS="$CFLAGS"
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 mkdir -p build && \
+if [ -d AppDir ];then
+	rm -rf AppDir
+fi
 mkdir -p AppDir/usr/bin && \
 mkdir -p AppDir/usr/lib && \
 cd build && \
@@ -25,7 +35,10 @@ if [ ! -d ezquake-source ];then
 	git clone https://github.com/ezQuake/ezquake-source.git
 fi
 cd ezquake-source && \
-git clean -qfdx
+if [ "$SKIP_BUILD" != "1" ];then
+	make clean
+	git clean -qfdx
+fi
 git reset --hard
 git pull
 if [ $? -ne 0 ];then
@@ -33,10 +46,10 @@ if [ $? -ne 0 ];then
 	exit 1
 fi
 REVISION=$(git log -n 1|head -1|awk '{print $2}'|cut -c1-6)
-make clean && \
 nice make -j$(nproc) && \
 cp -f ezquake-linux-x86_64 "$DIR/AppDir/usr/bin/." && \
-ln -sf usr/bin/ezquake-linux-x86_64 "$DIR/AppDir/AppRun" && \
+cp -f "$DIR/AppRun-x86_64" "$DIR/AppDir/AppRun" && \
+chmod +x "$DIR/AppDir/AppRun" && \
 echo "$DESKTOP_ENTRY" > "$DIR/AppDir/ezquake.desktop" && \
 cp "$DIR/quake.png" "$DIR/AppDir/."||true #copy over quake png if it exists
 cd "$DIR" && \
